more simplified collect speech tests
