+++ internal/e2e-realtime-api/src/utils.ts
@@ -43,7 +43,6 @@ interface CreateTestRunnerParams {
   testHandler: TestHandler
   executionTime?: number
   useDomainApp?: boolean
-  exitOnSuccess?: boolean
 }
 
 export const createTestRunner = ({
@@ -52,7 +51,6 @@ export const createTestRunner = ({
   testHandler,
   executionTime = MAX_EXECUTION_TIME,
   useDomainApp = false,
-  exitOnSuccess = true,
 }: CreateTestRunnerParams) => {
   let timer: ReturnType<typeof setTimeout>
 
@@ -67,9 +65,6 @@ export const createTestRunner = ({
     clearTimeout(timer)
     if (exitCode === 0) {
       console.log(`Test Runner ${name} Passed!`)
-      if (!exitOnSuccess) {
-        return
-      }
     } else {
       console.log(`Test Runner ${name} finished with exitCode: ${exitCode}`)
     }
